Seems like everything is working correctly now
